moved gpu test to the right file
